Move the scanning operation to another thread
