+++ gi_loadouts/face/scan/rule.py
@@ -1,3 +1,4 @@
+from PySide6.QtCore import QThread
 from PySide6.QtGui import QPixmap
 from PySide6.QtWidgets import QComboBox, QDialog, QLineEdit, QMessageBox
 
@@ -6,8 +7,8 @@ from gi_loadouts.data.arti import ArtiList
 from gi_loadouts.face.scan import areaiden
 from gi_loadouts.face.scan.file import file
 from gi_loadouts.face.scan.scan import Ui_scan
-from gi_loadouts.face.scan.util import modify_datatype_to_transfer, scan_artifact
-from gi_loadouts.face.util import truncate_text
+from gi_loadouts.face.scan.work import ScanWorker
+from gi_loadouts.face.util import modify_datatype_to_transfer, truncate_text
 from gi_loadouts.face.wind.talk import Dialog
 from gi_loadouts.type import arti
 from gi_loadouts.type.arti import ArtiLevl, __artistat__
@@ -29,7 +30,10 @@ class Rule(QDialog, Ui_scan, Dialog):
         self.shot = None
         self.snap = None
         self.part = ""
+        self.thread = None
+        self.worker = None
         self.dialog = QMessageBox()
+
         self.dist = {
             "Flower of Life": {"list": MainStatType_FWOL, "part": "fwol"},
             "Plume of Death": {"list": MainStatType_PMOD, "part": "pmod"},
@@ -168,6 +172,27 @@ class Rule(QDialog, Ui_scan, Dialog):
                 item.levl, item.rare, item.stat_name = levl.value.levl, rare.value.qant, stat
                 self.arti_data_main.setText(f"{round(item.stat_data, 1)}")
 
+    def kick_thread(self) -> None:
+        """
+        Kick things up after the thread has achieved its purpose
+
+        :return:
+        """
+        self.arti_text.setText("Inspecting screenshot...")
+        # self.worker.scan_artifact()
+
+    def wipe_thread(self, rslt) -> None:
+        """
+        Wipe things up after the thread has achieved its purpose
+
+        :return:
+        """
+        print("thread_cleanup INTO")
+        self.arti_text.setText("Browse your local storage to load a high quality screenshot of your artifact and the statistics will automatically be computed from there.")
+        # self.register_return_from_scanning(rslt)
+        self.thread.quit()
+        self.thread.wait()
+
     def load_reader(self) -> None:
         """
         Facilitate the loading of data and calculation of statistics
@@ -175,32 +200,45 @@ class Rule(QDialog, Ui_scan, Dialog):
         :return:
         """
         try:
-            self.arti_text.setText("Inspecting screenshot...")
             self.shot, self.snap = file.load_screenshot(self, "Select location to load artifact screenshot")
-            self.arti_shot.setPixmap(self.shot)
-            area, main, seco, team, levl, rare, duration = scan_artifact(self.snap)
-            if area in [self.arti_dist.itemText(indx) for indx in range(self.arti_dist.count())]:
-                self.arti_dist.setCurrentText(area)
-            if team in [self.arti_type.itemText(indx) for indx in range(self.arti_type.count())]:
-                self.arti_type.setCurrentText(team)
-            if rare in [self.arti_rare.itemText(indx) for indx in range(self.arti_rare.count())]:
-                self.arti_rare.setCurrentText(rare)
-            if levl in [self.arti_levl.itemText(indx) for indx in range(self.arti_levl.count())]:
-                self.arti_levl.setCurrentText(levl)
-            if main.stat_name.value in [self.arti_name_main.itemText(indx) for indx in range(self.arti_name_main.count())]:
-                self.arti_name_main.setCurrentText(main.stat_name.value)
-            for alfa, item in seco.items():
-                if item.stat_name.value in [getattr(self, f"arti_name_{alfa}").itemText(indx) for indx in range(getattr(self, f"arti_name_{alfa}").count())]:
-                    getattr(self, f"arti_name_{alfa}").setCurrentText(item.stat_name.value)
-                    getattr(self, f"arti_data_{alfa}").setText(str(item.stat_data))
+
+            self.thread, self.worker = QThread(), ScanWorker(self.snap)
+            self.worker.moveToThread(self.thread)
+            self.thread.started.connect(self.kick_thread)
+            self.thread.started.connect(self.worker.scan_artifact)
+            self.worker.result.connect(self.register_return_from_scanning)
+            self.worker.result.connect(self.wipe_thread)
+            self.thread.start()
         except Exception as expt:
             self.show_dialog(
                 QMessageBox.Information,
                 "Faulty scanning",
                 f"Please consider checking your input after ensuring that the proper Tesseract OCR executable has been selected.\n\n{expt}"
             )
-        finally:
-            self.arti_text.setText("Browse your local storage to load a high quality screenshot of your artifact and the statistics will automatically be computed from there.")
+
+    def register_return_from_scanning(self, rslt: tuple) -> None:
+        """
+        Place the selected screenshot and the detected attributes on the user interface
+
+        :return:
+        """
+        print("register_return_from_scanning", rslt)
+        self.arti_shot.setPixmap(self.shot)
+        area, main, seco, team, levl, rare, duration = rslt
+        if area in [self.arti_dist.itemText(indx) for indx in range(self.arti_dist.count())]:
+            self.arti_dist.setCurrentText(area)
+        if team in [self.arti_type.itemText(indx) for indx in range(self.arti_type.count())]:
+            self.arti_type.setCurrentText(team)
+        if rare in [self.arti_rare.itemText(indx) for indx in range(self.arti_rare.count())]:
+            self.arti_rare.setCurrentText(rare)
+        if levl in [self.arti_levl.itemText(indx) for indx in range(self.arti_levl.count())]:
+            self.arti_levl.setCurrentText(levl)
+        if main.stat_name.value in [self.arti_name_main.itemText(indx) for indx in range(self.arti_name_main.count())]:
+            self.arti_name_main.setCurrentText(main.stat_name.value)
+        for alfa, item in seco.items():
+            if item.stat_name.value in [getattr(self, f"arti_name_{alfa}").itemText(indx) for indx in range(getattr(self, f"arti_name_{alfa}").count())]:
+                getattr(self, f"arti_name_{alfa}").setCurrentText(item.stat_name.value)
+                getattr(self, f"arti_data_{alfa}").setText(str(item.stat_data))
 
     def load_tessexec(self) -> None:
         """
