+++ gi_loadouts/face/util.py
@@ -38,3 +38,14 @@ def modify_graphics_resource(path: str, radius: float = 2.5, shadow: float = 0.5
     except UnidentifiedImageError:
         rtrn = QPixmap(path)
     return rtrn
+
+
+def modify_datatype_to_transfer(text: str = ""):
+    if text.strip() == "":
+        rtrn = 0.0
+    else:
+        try:
+            rtrn = float(text)
+        except ValueError:
+            rtrn = 0.0
+    return rtrn
